Domain adjustments.

+++ src/main/java/com/medical/domain/Coordinates.java
@@ -1,26 +1,36 @@
 package com.medical.domain;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.maps.GeoApiContext;
+import com.google.maps.GeocodingApi;
+import com.google.maps.errors.ApiException;
+import com.google.maps.model.GeocodingResult;
+
 import javax.persistence.Embeddable;
+import java.io.IOException;
 import java.math.BigDecimal;
 
 @Embeddable
 public class Coordinates {
-    private BigDecimal x;
-    private BigDecimal y;
+    private double x;
+    private double y;
 
-    public BigDecimal getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(BigDecimal x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public BigDecimal getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(BigDecimal y) {
+    public void setY(double y) {
         this.y = y;
     }
+
+
 }
